restore event geometry after episode panel closed
