+++ src/core/focused_geometry/constants.ts
@@ -1,0 +1,1 @@
+export const FOCUSED_GEOMETRY_LOGICAL_LAYER_ID = 'focused-geometry';
