+++ src/core/focused_geometry/utils.ts
@@ -1,0 +1,7 @@
+import type { FocusedGeometry } from './types';
+
+export function getEventId(focusedGeometry: FocusedGeometry | null) {
+  return focusedGeometry?.source?.type === 'event'
+    ? focusedGeometry.source.meta.eventId
+    : null;
+}
