+++ src/features/analytics_panel/atoms/currentEventData.ts
@@ -1,4 +1,4 @@
-import { focusedGeometryAtom } from '~core/shared_state';
+import { focusedGeometryAtom } from '~core/focused_geometry/model';
 import { createAtom } from '~utils/atoms';
 import type { Event } from '~core/types';
 
@@ -9,10 +9,7 @@ export const currentEventDataAtom = createAtom(
   ({ get }, state: Event | null = null) => {
     const focusedGeometry = get('focusedGeometry');
     if (focusedGeometry) {
-      return (
-        ('meta' in focusedGeometry.source && focusedGeometry.source?.meta) ||
-        null
-      );
+      return ('meta' in focusedGeometry.source && focusedGeometry.source?.meta) || null;
     }
     return null;
   },
