+++ src/features/boundary_selector/atoms/boundaryMarkerAtom.ts
@@ -3,7 +3,8 @@ import { currentMapAtom, currentMapPositionAtom } from '~core/shared_state';
 import { appConfig } from '~core/app_config';
 import { constructOptionsFromBoundaries } from '~utils/map/boundaries';
 import { convertToAppMarker } from '~utils/map/markers';
-import { toolbarControlsAtom, focusedGeometryAtom } from '~core/shared_state';
+import { toolbarControlsAtom } from '~core/shared_state';
+import { focusedGeometryAtom } from '~core/focused_geometry/model';
 import { i18n } from '~core/localization';
 import { getCameraForGeometry } from '~utils/map/cameraForGeometry';
 import { BOUNDARY_MARKER_ID } from '../constants';
