+++ src/features/focused_geometry_layer/index.ts
@@ -4,7 +4,7 @@ import { layersSettingsAtom } from '~core/logical_layers/atoms/layersSettings';
 import { createAsyncWrapper } from '~utils/atoms/createAsyncWrapper';
 import { store } from '~core/store/store';
 import { forceRun } from '~utils/atoms/forceRun';
-import { FOCUSED_GEOMETRY_LOGICAL_LAYER_ID } from '~core/shared_state/focusedGeometry';
+import { FOCUSED_GEOMETRY_LOGICAL_LAYER_ID } from '~core/focused_geometry/constants';
 import {
   FOCUSED_GEOMETRY_LOGICAL_LAYER_TRANSLATION_KEY,
   FOCUSED_GEOMETRY_COLOR,
