+++ src/features/layers_in_area/atoms/areaLayersLegendsAndSources.ts
@@ -2,8 +2,8 @@ import { createAtom } from '~utils/atoms/createPrimitives';
 import { layersSourcesAtom } from '~core/logical_layers/atoms/layersSources';
 import { layersLegendsAtom } from '~core/logical_layers/atoms/layersLegends';
 import { legendFormatter } from '~features/layers_in_area/utils/legendFormatter';
-import { focusedGeometryAtom } from '~core/shared_state';
-import { getEventId } from '~core/shared_state/focusedGeometry';
+import { focusedGeometryAtom } from '~core/focused_geometry/model';
+import { getEventId } from '~core/focused_geometry/utils';
 import { areaLayersDetailsResourceAtom } from './areaLayersDetailsResource';
 import { areaLayersDetailsResourceAtomCache } from './areaLayersDetailsResource/areaLayersDetailsResourceAtomCache';
 import type { Action } from '@reatom/core';
